moves kafkametrics dd implementation to package
